Hack to get erb preprocessor working on Windows
The problem: rubyinstaller on Windows makes an "erb" and "erb.bat",
but no "erb.exe". The win32 function readProcess ends up using to
locate an executable therefore does not find it. callCommand
passes a string directly to the shell, and fortunately the same
syntax works on linux and windows (both cmd.exe and mingw).